refactor: 예약, 대기 삭제 로직에 논리 트랜잭션 적용
